xiaozhi-client/xiaozhi-web-client/pyproject.toml
[tool.poetry]
name = "xiaozhi-web-client-main"
version = "0.1.0"
description = ""
authors = ["kalicyh <34980061+kaliCYH@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
websockets = "12.0"
flask = "3.0.2"
python-dotenv = "1.0.0"
opuslib = "3.0.1"
numpy = "1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
